Separates process_record_kvm into separate function in new processs_records file.
