+++ keyboards/asedlack/macrostick/keymaps/default/keymap.c
@@ -1,9 +1,5 @@
 #include QMK_KEYBOARD_H
-
-enum custom_keycodes {
-  KVM_1 = SAFE_RANGE,
-  KVM_2,
-};
+#include "../../process_records.c"
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
   [0] = LAYOUT(
@@ -17,23 +13,9 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 
 // Process macros
 bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-  switch (keycode) {
-    case KVM_1:
-      if (record->event.pressed) {
-        tap_code(KC_SCRL);
-        tap_code(KC_SCRL);
-        tap_code(KC_1);
-      }
-      break;
-    case KVM_2:
-      if (record->event.pressed) {
-        tap_code(KC_SCRL);
-        tap_code(KC_SCRL);
-        tap_code(KC_2);
-      }
-      break;
-  }
-  return true;
+  return (
+       process_record_kvm(keycode, record)
+  );
 };
 
 // Process encoders
